+++ src/scripts/events/detectPageLoad.tab.js
