+++ src/scripts/events/notify.bg.js
