+++ src/scripts/events/uiEnhancer.tab.js
